Improved some behaviors
-add avoid obstacle against polygon,
BarrierAvoidanceWithThreeFeelersPolygon
-set "Name" variable to all behaviors.
-change List<Circle> to PositionedObjectList<Circle>, and other type
like AAR and Polygon.  The reason to do this is it will make it easier
to call.
Before: avoidO.CircleObstacles =
WorldCollision.Circles.ToList<Circle>();
After the change: avoidO.CircleObstacles = WorldCollision.Circles;
-can now call .Add(circle or AARect or Polygon) for
AvoidObstaclesBehavior.
e.g, avoidO.CircleObstacles.Add(CollisionCircle);
-added StopDistance variable to Seek and GoTo related Behaviors.  When
entity is within the StopDistance of target, then it will stop.
before it must be stopped at the destination x and y (right on the dot)
now it checks if it's in range specified, then stop.
this should prevent non stopping when force or velocity speed is high,
otherwise the entity may keep circling around the target unable to stop